+++ src/app/(dashboard)/plan/initial/[id]/page.tsx
@@ -1,0 +1,7 @@
+import PlanInitialForm from '@/components/PlanInitialForm';
+
+function Plan({ params }: { params: { id: string } }) {
+  return <PlanInitialForm articlePageId={params.id} />;
+}
+
+export default Plan;
